refactor: move command handler to separate file, move invite msg to utils
